Delete Nea. population right after admixture
This simplifies parsing the SLiM output file when we want to analyze the
distribution of selection coefficients of introgressed Nea. alleles and
their frequencies in the modern human population. Since the only
population in the simulation will be the modern human one, the
frequencies of all mutations in the output correspond to the frequencies
in modern humans (and note modern humans and Neanderthals together).
This saves us time of parsing the genomes of each individual to scan
for introgressed Nea. mutations to calculate their frequencies.